+++ Resources/model/facebookFriend.js
@@ -33,10 +33,15 @@ exports.getNameByFacebookId = function(_fbId){
 
 exports.populateFacebookFriend = function(_friendsCollection) {
 	Ti.API.info('populating facebook friend...');
+	if(_friendsCollection.length === 0) 
+		return;
+
 	var db = Ti.Database.open(Ti.App.DATABASE_NAME); 
 	db.execute('DELETE FROM FacebookFriend');
 	for(var i = 0; i < _friendsCollection.length; i++) {
+
 		var curFriend = _friendsCollection[i];
+		Ti.API.info('inserting fbFriend: '+JSON.stringify(curFriend));
 		var insertString = "INSERT INTO FacebookFriend(Id, FacebookId, Name, PictureUrl, City, IsInvited, ClosenessScore) VALUES(NULL,?,?,?,?,0,0)";
 		db.execute(insertString, curFriend.uid+"", curFriend.name, curFriend.pic_square, curFriend.city); //+"" for converting long to str
 	}
